requirements.txt
aiogram==3.15.0
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers==4.45.0
accelerate==1.0.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0
pandas==2.2.3
nltk==3.9.1
python-dotenv==1.0.1
langchain==1.2.10
langchain-community==0.4.1
langchain-huggingface==1.2.1
unsloth
nest-asyncio